refactor(redis_ops): extract commands module
Extract command generation functions into dedicated commands.rs module:
- generate_string_command: Generates SET/SETEX commands
- generate_list_commands: Generates RPUSH commands
- generate_set_commands: Generates SADD commands
- generate_zset_commands: Generates ZADD commands
- generate_hash_commands: Generates HMSET commands

All 3 command generation tests moved with the functions.
